Create exact copy of block device mapping
- on replace-vm
- remove filtering of root device
- copy snapshot id
- resolve conflicts in encryption and snapshot id fields

[#148434573]

+++ client_test.go
@@ -27,7 +27,7 @@ var _ = Describe("test for unexported features", func() {
 					{
 						DeviceName: "/dev/sda1",
 						EBS: EBS{
-							VolumeSize: int64(50),
+							VolumeSize: 50,
 						},
 					},
 				},
